LayoutTests/imported/w3c/web-platform-tests/speculation-rules/speculation-tags/cross-site-to-same-site-redirection-prefetch.https-expected.txt


FAIL same-site redirection prefetch should have Sec-Speculation-Tags even if cross-site url in the redirection chain assert_equals: expected "\"tag1\"" but got "tag1"
